backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.135.3
fastapi-cli==0.0.24
h11==0.16.0
httptools==0.7.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
rich==15.0.0
rich-toolkit==0.19.7
shellingham==1.5.4
starlette==1.0.0
tomli==2.4.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# HTTP client
httpx==0.27.0

# Requests (WhatsApp integration)
requests==2.31.0

# Email validation (pydantic extra)
email-validator==2.1.1
